fix previous tests
